+++ solstice/src/main/java/dev/equo/solstice/ShimStorage.java
@@ -16,18 +16,18 @@ package dev.equo.solstice;
 import java.io.File;
 import java.util.Map;
 import javax.annotation.Nullable;
-import org.eclipse.osgi.service.datalocation.Location;
+import org.eclipse.osgi.internal.location.EquinoxLocations;
 import org.slf4j.Logger;
 
 class ShimStorage {
 	private final File configDir;
 
 	ShimStorage(Map<String, String> props, Logger logger) {
-		String configArea = props.get(Location.CONFIGURATION_AREA_TYPE);
+		String configArea = props.get(EquinoxLocations.PROP_CONFIG_AREA);
 		if (configArea == null) {
 			logger.warn(
 					"Recommend setting "
-							+ Location.CONFIGURATION_AREA_TYPE
+							+ EquinoxLocations.PROP_CONFIG_AREA
 							+ " to a directory, getDataFile will return null");
 			this.configDir = null;
 		} else {