+++ solstice/CHANGELOG.md
@@ -5,6 +5,7 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format.
 ## [Unreleased]
 ### Fixed
 - Handle `system.bundle` in `Require-Bundle`.
+- `ShimIdeBootstrapServices` is now compatible with older Eclipse versions.
 
 ## [0.18.0] - 2023-03-10
 ### Added